release 1.8.0

See merge request vng/diwi!780

+++ backend/src/main/java/nl/vng/diwi/models/PlanModel.java
@@ -369,7 +369,7 @@ public class PlanModel extends DatedDataModelSuperClass {
     public static class PlanOwnershipModel {
 
         private OwnershipType type;
-        private SingleValueOrRangeModel<Integer> value;
+        private SingleValueOrRangeModel<Long> value;
         private SelectModel rangeCategoryOption;
 
         public PlanOwnershipModel(PlanSqlModel.PlanSqlOwnershipOptionModel sqlOption) {
